+++ tests/easy_uds_test/common.hpp
@@ -103,5 +103,6 @@ void test_closing_connection_counts_toward_limit();
 void test_stop_interrupts_blocked_workers();
 void test_handler_error_opt_out();
 void test_global_memory_budgets();
+void test_partial_request_uses_global_budget();
 
 } // namespace easy_uds::test